pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agrifield"
version = "0.1.0"
description = "Agricultural field data analytics — acre-level crop yield prediction"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Proffitt Blackburn" },
]
dependencies = [
    "click>=8.0",
    "geopandas>=0.14",
    "matplotlib>=3.7",
    "numpy>=1.24",
    "pandas>=2.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyproj>=3.5",
    "pyyaml>=6.0",
    "requests>=2.28",
    "scikit-learn>=1.3",
    "shapely>=2.0",
    "tqdm>=4.60",
    "contextily>=1.3",
]

[project.optional-dependencies]
catboost = ["catboost>=1.2"]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]
all = [
    "agrifield[catboost]",
    "agrifield[dev]",
]

[project.scripts]
agrifield = "agrifield.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/agrifield"]

[tool.pytest.ini_options]
testpaths = ["tests"]
